moved ips to isotopes folder, organized packages
